+++ CHANGELOG.md
@@ -12,10 +12,11 @@ Most recent releases are shown at the top. Each release shows:
 - N/A
 
 ### changed
+- Added `prompt_template` parameter to `LLM` constructor
 - Added `update_max_tokens` and `update_stop` methods to `LLM` for dynamic adjustments during prompt experiments
 
 ### fixed:
-- N/A
+- Explicitly set `offload_kqv` to ensure GPUs are fully utilized
 
 
 ## 0.0.32 (2023-12-10)